solved all the problem
