+++ .eslintrc.json
@@ -14,13 +14,12 @@
     "no-shadow": "off",
     "no-param-reassign": "off",
     "eol-last": "off",
-    "import/extensions": [
-      1,
-      {
-        "js": "always",
-        "json": "always"
-      }
-    ]
+    "import/extensions": [ 1, {
+      "js": "always", "json": "always"
+    }]
   },
-  "ignorePatterns": ["dist/", "build/"]
+  "ignorePatterns": [
+    "dist/",
+    "build/"
+  ]
 }
