+++ index.html
@@ -80,12 +80,12 @@
         <li>Our Phone Number: 003311223344</li>
         <li>Our Adress: John Doe street, 22, 883322 Vikings, Valhala</li>
       </ul>
-     </section>
+    </section>
     <!-- Footer Section -->
     <footer><p>Copyright: All rights reserved</p></footer>
     </main>
-   
-     <script  type="module" src="index.js"></script> 
+  
+    <script  type="module" src="index.js"></script> 
     <script src="https://cdn.jsdelivr.net/npm/luxon@1.25.0/build/global/luxon.min.js"></script>
   </body>
 </html>
