+++ package.json
@@ -4,10 +4,32 @@
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.26.0",
-    "hint": "^6.1.9",
+    "hint": "^7.1.3",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
     "stylelint-csstree-validator": "^1.9.0",
     "stylelint-scss": "^3.21.0"
-  }
+  },
+  "dependencies": {
+    "luxon": "^3.1.0",
+    "three": "^0.146.0"
+  },
+  "name": "awesome_books_with_es6",
+  "description": "![](https://img.shields.io/badge/Microverse-blueviolet)",
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ZuhRa-HashiMi/Awesome_books_with_ES6.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/ZuhRa-HashiMi/Awesome_books_with_ES6/issues"
+  },
+  "homepage": "https://github.com/ZuhRa-HashiMi/Awesome_books_with_ES6#readme"
 }
